chore: Player and Card models

+++ app/models/card.py
@@ -1,5 +1,7 @@
 class Card:
-    def __init__(self, name: str, age: int, size: float, height: float, weight: float):
+    def __init__(
+        self, name: str, age: float, size: float, height: float, weight: float
+    ):
         self.name = name
         self.age = age
         self.size = size
@@ -10,4 +12,3 @@ class Card:
         return "name: {0}, age: {1}, size: {2}, height: {3}, weight: {4}".format(
             self.name, self.age, self.size, self.height, self.weight
         )
-
